[refactor] (doge) move common HD test to helper.ts
