+++ packages/coinlib-doge/test/fixtures/accounts.ts
@@ -4,18 +4,26 @@ import { ROOT_XPRV, MNEMONIC } from './bip44'
 // Values calculated using https://iancoleman.io/bip39/
 
 export type AccountFixture = {
-  derivationPath: string,
-  xprv: string,
-  xpub: string,
-  sweepTxSize: number,
+  derivationPath: string
+  xprv: string
+  xpub: string
+  sweepTxSize: number
   addresses: {
-    [i: number]: string,
+    [i: number]: string
+  }
+  balances: {
+    [i: number]: {
+      confirmedBalance: string
+      unconfirmedBalance: string
+      spendableBalance: string
+      sweepable: boolean
+    }
   }
 }
 
 export const legacyAccount: AccountFixture = {
   derivationPath: "m/44'/3'/0'",
-  xprv: ROOT_XPRV,  // ??
+  xprv: ROOT_XPRV, // ??
   sweepTxSize: 192,
   xpub:
     'xpub6CHVE7XJJHooX3pFQSxi733aXzjeLgbVMimR3kvTfYx1cPtBmEaJT6bn3CJDhdysxbYes8yH1yNEX8dJxSTBxgBD6GUmsGrzGmsAKEax96X',
@@ -27,7 +35,21 @@ export const legacyAccount: AccountFixture = {
     '6': 'DQZb8FC9PpqMZfVFwVR6UnCB5YgT67D4Vz',
     '10': 'DC2uUxHtZdSAQ67WgZ7rMw1rKFAheKBu7F',
     '10000': 'DQuTrS22J4yCHRotbSNEnE1gnwBaR7YjN6',
-    '20000': 'DGEgJ5j44HWG1sXgsn7zS1Yi5DPzQXbMeq'
+    '20000': 'DGEgJ5j44HWG1sXgsn7zS1Yi5DPzQXbMeq',
+  },
+  balances: {
+    '0': {
+      confirmedBalance: '993.02',
+      unconfirmedBalance: '0',
+      spendableBalance: '993.02',
+      sweepable: true,
+    },
+    '1': {
+      confirmedBalance: '300',
+      unconfirmedBalance: '0',
+      spendableBalance: '300',
+      sweepable: true,
+    },
   },
 }
 
@@ -36,18 +58,30 @@ export const accountsByAddressType: { [type in SinglesigAddressType]: AccountFix
 }
 
 export type SeedAccountFixture = {
-  seed: string,
-  uniPubKey: string,
-  derivationPath: string,
-  sweepTxSize: number,
+  seed: string
+  // uniPubKey: string
+  xpub: string
+  derivationPath: string
+  sweepTxSize: number
   addresses: {
-    [i: number]: string,
+    [i: number]: string
+  }
+  balances: {
+    [i: number]: {
+      confirmedBalance: string
+      unconfirmedBalance: string
+      spendableBalance: string
+      sweepable: boolean
+    }
   }
 }
 
 export const seedLegacyAccount: SeedAccountFixture = {
   seed: MNEMONIC,
-  uniPubKey: 'xpub6D5k1geCx56YrTHmCB51B6opJasQ8dqpPSy1cdYF9wDYPRFFDLSnDQCzqfXEHSYsCS4LJerruc5Rdd5uL8gPEkHu4UxCVtoa7xjgzWhRgkf',
+  // uniPubKey:
+  // 'xpub6D5k1geCx56YrTHmCB51B6opJasQ8dqpPSy1cdYF9wDYPRFFDLSnDQCzqfXEHSYsCS4LJerruc5Rdd5uL8gPEkHu4UxCVtoa7xjgzWhRgkf',
+  xpub:
+    'xpub6D5k1geCx56YrTHmCB51B6opJasQ8dqpPSy1cdYF9wDYPRFFDLSnDQCzqfXEHSYsCS4LJerruc5Rdd5uL8gPEkHu4UxCVtoa7xjgzWhRgkf',
   derivationPath: "m/44'/3'/0'",
   sweepTxSize: 192,
   addresses: {
@@ -58,12 +92,26 @@ export const seedLegacyAccount: SeedAccountFixture = {
     '6': 'DBA6mNMJ9pgtA6SkNtKvGEZZdf3g77VpTh',
     '10': 'DA97SgA5chqPDYgj46rMFxLR65HWdxYWiB',
     '10000': 'DAZnRgdk6AWzUPcAKo26qNcnVqQ2xuESsF',
-    '20000': 'D8BxgjtuYMfH5tsAACBjeQwAwoFYYE4jy9'
+    '20000': 'D8BxgjtuYMfH5tsAACBjeQwAwoFYYE4jy9',
+  },
+  balances: {
+    '0': {
+      confirmedBalance: '0',
+      unconfirmedBalance: '0',
+      spendableBalance: '0',
+      sweepable: false,
+    },
+    '1': {
+      confirmedBalance: '0',
+      unconfirmedBalance: '0',
+      spendableBalance: '0',
+      sweepable: false,
+    },
   },
 }
 
-export const seedLegacyAccountXPrv: string = 'xprv9s21ZrQH143K3AVAFiWZgcBVo4CYcuyaLwXNQvefrNjqdUF6WwpXVYrHhmGYJXD5hzuTpvvyRxCfUxEUeRkTBgj1CzsvHBdtheNnBXpuu8p'
-
+export const seedLegacyAccountXPrv: string =
+  'xprv9s21ZrQH143K3AVAFiWZgcBVo4CYcuyaLwXNQvefrNjqdUF6WwpXVYrHhmGYJXD5hzuTpvvyRxCfUxEUeRkTBgj1CzsvHBdtheNnBXpuu8p'
 
 export const seedAccountsByAddressType: { [type in SinglesigAddressType]: SeedAccountFixture } = {
   [AddressType.Legacy]: seedLegacyAccount,
